Refactor: Tags 관련 컴포넌트 일체화
